+++ cards/src/test/java/com/hiddenswitch/spellsource/tests/cards/ArchaeologistTests.java
@@ -92,4 +92,14 @@ public class ArchaeologistTests extends TestBase {
 			assertEquals(3, player.getHand().size());
 		});
 	}
+
+	@Test
+	public void testGoldRush() {
+		runGym((context, player, opponent) -> {
+			playCard(context, player, "spell_gold_rush");
+			for (int i = 0; i < 8; i++) {
+				assertEquals(0, costOf(context, player, player.getDeck().get(i)));
+			}
+		});
+	}
 }